% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFfmDT.R
\name{print.ffmSpec}
\alias{print.ffmSpec}
\title{print.ffmSpec}
\usage{
\method{print}{ffmSpec}(x, ...)
}
\arguments{
\item{x}{an object of class ffmSpec}

\item{...}{any other option}
}
\value{
No return value, called for displaying attributes
}
\description{
print.ffmSpec
}
