% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facmodCS-package.R
\docType{package}
\name{facmodCS-package}
\alias{facmodCS}
\alias{facmodCS-package}
\title{facmodCS: Cross-Section Factor Models}
\description{
Linear cross-section factor model fitting with least-squares and and robust fitting the lmrobdetMM() function from RobStatTM; related volatility, VaR and ES risk and performance attribution (factor-contributed vs idiosyncratic returns); tabular displays of risk and performance reports; factor model Monte Carlo. The package authors would like to thank Chicago Research on Security Prices,LLC (CRSP) for the cross-section of about 300 CRSP stocks data (in the data.table object stocksCRSP, and S&P GLOBAL MARKET INTELLIGENCE for contributing 14 factor scores (a.k.a "alpha factors".and "factor exposures") fundamental data on the 300 companies in the data.table object factorSPGMI. The stocksCRSP and factorsSPGMI data are not covered by the GPL-2 license, are not provided as open source pf any king, and they are not to be redistributed in any form.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robustport/facmodCS}
}

}
\keyword{internal}
